# runtime/internal/agentapi/openapi.yaml
openapi: 3.1.0
info:
  title: Signal Foundry agent runtime API
  description: |
    HTTP API for agent runs backed by the ADK `AgentRunner` / `Run` flow in this module.
    The runner uses `StreamingModeSSE`; HTTP responses use **Server-Sent Events** (`text/event-stream`).

    **Stream contract:** Each SSE message uses standard framing (`event:` optional, `data:` JSON).
    Every `data` line is a JSON object with an `event` discriminator (see `StreamEvent`).

    **Start vs continue:** `POST /agent-runs` begins a new session; the server assigns `sessionId`
    and emits it in the first `sessionBound` event. `POST /sessions/{sessionId}/agent-runs`
    continues an existing session (same request body shape; session comes from the path).
  version: 0.1.0
servers:
  - url: /
    description: Relative to the HTTP server base URL
...
paths:
...
  /models:
    get:
      tags: [Models]
      operationId: listModels
      summary: List all available models across all providers
      description: |
        Returns all models configured across all providers. Each entry includes the provider name
        to allow constructing the fully-qualified model name (`provider/model-name`) for agent runs.
      responses:
        '200':
          description: List of available models.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /providers:
    get:
      tags: [Providers]
      operationId: listProviders
      summary: List all provider configurations
      responses:
        '200':
          description: List of provider configurations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Providers]
      operationId: createProvider
      summary: Create a new provider configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderRequest'
      responses:
        '201':
          description: Provider configuration created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /providers/{providerName}:
    get:
      tags: [Providers]
      operationId: getProvider
      summary: Get a provider configuration by name
      parameters:
        - $ref: '#/components/parameters/ProviderName'
      responses:
        '200':
          description: Provider configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Providers]
      operationId: updateProvider
      summary: Update a provider configuration by name
      parameters:
        - $ref: '#/components/parameters/ProviderName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProviderRequest'
      responses:
        '200':
          description: Updated provider configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Providers]
      operationId: deleteProvider
      summary: Delete a provider configuration by name
      parameters:
        - $ref: '#/components/parameters/ProviderName'
      responses:
        '204':
          description: Provider configuration deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /agent-profiles:
    get:
      tags: [Agent profiles]
      operationId: listAgentProfiles
      summary: List all saved agent profiles
      responses:
        '200':
          description: List of saved agent profiles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentProfileListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Agent profiles]
      operationId: createAgentProfile
      summary: Create a new saved agent profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentProfileRequest'
      responses:
        '201':
          description: Agent profile created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /agent-profiles/{profileName}:
    get:
      tags: [Agent profiles]
      operationId: getAgentProfile
      summary: Get a saved agent profile by name
      parameters:
        - $ref: '#/components/parameters/ProfileName'
      responses:
        '200':
          description: Agent profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Agent profiles]
      operationId: updateAgentProfile
      summary: Update mutable fields for a saved agent profile
      parameters:
        - $ref: '#/components/parameters/ProfileName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentProfileRequest'
      responses:
        '200':
          description: Updated agent profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Agent profiles]
      operationId: deleteAgentProfile
      summary: Delete a saved agent profile by name
      parameters:
        - $ref: '#/components/parameters/ProfileName'
      responses:
        '204':
          description: Agent profile deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
...
    ProviderName:
      name: providerName
      in: path
      required: true
      description: Unique provider name (primary key, e.g. `openai`, `openrouter`).
      schema:
        type: string
        minLength: 1
    ProfileName:
      name: profileName
      in: path
      required: true
      description: Unique profile name (primary key, e.g. `coding-agent`).
      schema:
        type: string
        minLength: 1
  schemas:
...
    ModelConfig:
      type: object
      description: A model available through a provider.
      required: [name]
      properties:
        name:
          type: string
          description: Technical model identifier (e.g., "gpt-4.1").
          minLength: 1
        displayName:
          type: string
          description: Optional human-friendly label (e.g., "GPT 4.1").
        summarization:
          type: boolean
          description: >
            Designates this model for summarization tasks (e.g. generating session titles).
            If multiple models have this enabled, the first one found is used.
          default: false
      additionalProperties: false
    ModelInfo:
      type: object
      description: A model available across providers, including the provider name for fully-qualified identification.
      required: [provider, name]
      properties:
        provider:
          type: string
          description: Provider name (prefix in fully-qualified model names like `provider/model-name`).
        name:
          type: string
          description: Technical model identifier (e.g., "gpt-4.1").
        displayName:
          type: string
          description: Optional human-friendly label.
      additionalProperties: false
    ModelListResponse:
      type: object
      description: List of available models across all providers.
      required: [models]
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelInfo'
      additionalProperties: false
...
    CreateProviderRequest:
      type: object
      description: Request body for creating a new provider configuration.
      required: [name, type, baseUrl, apiKey]
      properties:
        name:
          type: string
          description: Unique technical name (primary key). Must match `^[a-z][a-z0-9-]*$`. Immutable after creation.
          minLength: 1
        type:
          type: string
          description: Provider protocol type. Only `openai-compatible` is supported. Immutable after creation.
          minLength: 1
        displayName:
          type: string
          description: Optional human-friendly label.
        baseUrl:
          type: string
          description: Base URL of the provider API endpoint.
          minLength: 1
        apiKey:
          type: string
          description: API key for authentication. Stored in plaintext (local/dev use).
          minLength: 1
        models:
          type: array
          description: Models available through this provider.
          items:
            $ref: '#/components/schemas/ModelConfig'
      additionalProperties: false
    UpdateProviderRequest:
      type: object
      description: Request body for updating a provider configuration. `name` and `type` are immutable and not accepted here.
      required: [baseUrl]
      properties:
        displayName:
          type: string
          description: Human-friendly label.
        baseUrl:
          type: string
          description: Base URL of the provider API endpoint.
          minLength: 1
        apiKey:
          type: string
          description: API key for authentication. Omit to keep the current value.
        models:
          type: array
          description: Models available through this provider. Replaces the existing list.
          items:
            $ref: '#/components/schemas/ModelConfig'
      additionalProperties: false
    ProviderResponse:
      type: object
      description: A provider configuration as returned by the API (API key is masked).
      required: [name, type, baseUrl, apiKeyPreview, createdAt, updatedAt, models]
      properties:
        name:
          type: string
          description: Unique technical name (primary key).
        type:
          type: string
          description: Provider protocol type.
        displayName:
          type: string
          description: Human-friendly label (may be empty).
        baseUrl:
          type: string
          description: Base URL of the provider API endpoint.
        apiKeyPreview:
          type: string
          description: Masked API key preview (e.g. `...xYz1`). Never the full key.
        models:
          type: array
          description: Models available through this provider.
          items:
            $ref: '#/components/schemas/ModelConfig'
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last-update timestamp.
      additionalProperties: false
    ProviderListResponse:
      type: object
      description: List of provider configurations.
      required: [providers]
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderResponse'
      additionalProperties: false
    AgentProfileExecutionSettings:
      oneOf:
        - $ref: '#/components/schemas/AgentProfileRegularExecutionSettings'
        - $ref: '#/components/schemas/AgentProfileACPStdioExecutionSettings'
    AgentProfileRegularExecutionSettings:
      type: object
      description: Runtime-owned execution defaults for built-in profile execution.
      required: [defaultModel]
      properties:
        mode:
          type: string
          enum: [regular]
          description: Execution mode. Omit for regular built-in execution.
        defaultModel:
          type: string
          description: Fully-qualified default model in `provider/model-name` form.
          minLength: 1
      additionalProperties: false
    AgentProfileACPStdioExecutionSettings:
      type: object
      description: Runtime-owned execution defaults for ACP-compatible stdio profile execution.
      required: [mode, agentCommand]
      properties:
        mode:
          type: string
          const: acp-stdio
          description: Execution mode for ACP-compatible stdio profiles.
        agentCommand:
          $ref: '#/components/schemas/ACPStdioAgentCommand'
        cwd:
          type: string
          description: Optional working directory used to launch the ACP stdio agent.
      additionalProperties: false
    ACPStdioAgentCommand:
      type: object
      description: Command defaults used to launch an ACP-compatible stdio agent.
      required: [command]
      properties:
        command:
          type: string
          minLength: 1
          description: Command executable to launch the ACP-compatible stdio agent.
        args:
          type: array
          description: Command arguments passed to the ACP-compatible stdio agent executable.
          uniqueItems: true
          items:
            type: string
            minLength: 1
      additionalProperties: false
    CreateAgentProfileRequest:
      type: object
      description: Request body for creating a saved agent profile.
      required: [name, role, instructions, executionSettings]
      properties:
        name:
          type: string
          description: Unique technical profile name (primary key). Immutable after creation.
          minLength: 1
        displayName:
          type: string
          description: Optional human-friendly profile label.
        role:
          type: string
          description: Profile role describing high-level responsibility.
          minLength: 1
        instructions:
          type: string
          description: Profile-specific behavior instructions.
          minLength: 1
        toolRefs:
          type: array
          description: Tool references available to this profile.
          items:
            type: string
            minLength: 1
        executionSettings:
          $ref: '#/components/schemas/AgentProfileExecutionSettings'
      additionalProperties: false
    UpdateAgentProfileRequest:
      type: object
      description: Request body for updating a saved agent profile. `name` is immutable and not accepted here.
      required: [role, instructions, executionSettings]
      properties:
        displayName:
          type: string
          description: Human-friendly profile label.
        role:
          type: string
          description: Profile role describing high-level responsibility.
          minLength: 1
        instructions:
          type: string
          description: Profile-specific behavior instructions.
          minLength: 1
        toolRefs:
          type: array
          description: Tool references available to this profile. Replaces the existing list.
          items:
            type: string
            minLength: 1
        executionSettings:
          $ref: '#/components/schemas/AgentProfileExecutionSettings'
      additionalProperties: false
    AgentProfileResponse:
      type: object
      description: A saved agent profile as returned by the API.
      required:
        - name
        - displayName
        - role
        - instructions
        - toolRefs
        - executionSettings
        - createdAt
        - updatedAt
      properties:
        name:
          type: string
          description: Unique technical profile name (primary key).
        displayName:
          type: string
          description: Human-friendly profile label (may be empty).
        role:
          type: string
          description: Profile role describing high-level responsibility.
        instructions:
          type: string
          description: Profile-specific behavior instructions.
        toolRefs:
          type: array
          description: Tool references available to this profile.
          items:
            type: string
        executionSettings:
          $ref: '#/components/schemas/AgentProfileExecutionSettings'
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last-update timestamp.
      additionalProperties: false
    AgentProfileListResponse:
      type: object
      description: List of saved agent profiles.
      required: [profiles]
      properties:
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/AgentProfileResponse'
      additionalProperties: false
    ProblemDetails:
      type: object
      description: RFC 9457 problem details (subset).
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      additionalProperties: true
  responses:
    BadRequest:
      description: Malformed request body or validation failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Authentication required or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Resource name conflict (duplicate key).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'